Refactor password generator, rename method in generator service
