+++ Api/Services/Management/IApiClientManagementService.cs
@@ -1,6 +1,5 @@
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Controllers.AdministratorControllers;
 using HappyTravel.Edo.Api.Models.Agents;
 
 namespace HappyTravel.Edo.Api.Services.Management
@@ -9,6 +8,6 @@ namespace HappyTravel.Edo.Api.Services.Management
     {
         Task<Result> Set(int agencyId, int agentId, ApiClientData clientData);
         Task<Result> Delete(int agencyId, int agentId);
-        Task<Result<ApiClientData>> GenerateApiClient(int agencyId, int agentId);
+        Task<Result<ApiClientData>> Generate(int agencyId, int agentId);
     }
 }
